+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Scheduled callback interface and interval implementation to the standard component implementation.
 - `always_defer` option to slash commands (not including groups).
+- `tanjun.MessageCommand` is now callable like `tanjun.SlashCommand`.
 
 ## [2.2.3a1] - 2022-01-06
 ### Added